detection-consumer/requirements.txt
elasticsearch==8.16.0
kafka-python==2.0.2
minio==7.2.11
ultralytics==8.3.34
